- Win: mforms: fixed var type in enumeration (signed vs unsigned) which caused a cast exception. - Fixes MW-126 [MODEL] Reverse engineer error when moving one table between lists
[TESTS]
Started creation of a model from a schema, opened the filters in the object selection dialog and moved entries from one list to the other (single, multiple, all, both directions). The list of selected indices is currently only used in object filters.

+++ library/forms/winforms/src/wf_listbox.cpp
@@ -159,7 +159,7 @@ std::vector<size_t> ListBoxWrapper::get_selected_indices(mforms::ListBox *backen
 {
   std::vector<size_t> result;
   ListBox ^listbox = ListBoxWrapper::GetManagedObject<ListBox>(backend);
-  for each (size_t index in listbox->SelectedIndices)
+  for each (int index in listbox->SelectedIndices) // It's an arrow of Int32. Don't change to size_t.
     result.push_back(index);
   return result;
 }
